Move mypyc.genopsvtable to mypyc.irbuild.vtable
