+++ mypyc/genopsmain.py
@@ -31,7 +31,7 @@ from mypyc.errors import Errors
 from mypyc.options import CompilerOptions
 from mypyc.ops import ModuleIR, ModuleIRs
 from mypyc.prebuildvisitor import PreBuildVisitor
-from mypyc.genopsvtable import compute_vtable
+from mypyc.irbuild.vtable import compute_vtable
 from mypyc.irbuild.prepare import build_type_map
 from mypyc.genops import IRBuilder
 from mypyc.genopsvisitor import IRBuilderVisitor
