+++ mypyc/irbuild/vtable.py
